+++ packages/api/src/object/Attachment.ts
@@ -25,12 +25,10 @@ export interface Attachment {
    */
   fetchContents(): Promise<Response>;
 }
-/**
- * This interface should also accept the File object from
- * the W3C FileApi https://www.w3.org/TR/FileAPI/#file-section
- */
-export interface AttachmentUpload extends Blob {
+
+export interface AttachmentUpload {
   readonly name: string;
+  readonly data: Blob;
 }
 
 export interface AttachmentMetadata {