+++ packages/client/src/actions/actions.test.ts
@@ -217,10 +217,24 @@ describe("actions", () => {
       typeof clientBoundBatchActionTakesAttachment
     >[0];
 
-    expectTypeOf<{ attachment: string | AttachmentUpload }>().toMatchTypeOf<
+    expectTypeOf<
+      {
+        attachment:
+          | string
+          | AttachmentUpload
+          | Blob & { readonly name: string };
+      }
+    >().toMatchTypeOf<
       InferredParamType
     >();
-    expectTypeOf<{ attachment: string | AttachmentUpload }[]>().toMatchTypeOf<
+    expectTypeOf<
+      {
+        attachment:
+          | string
+          | AttachmentUpload
+          | Blob & { readonly name: string };
+      }[]
+    >().toMatchTypeOf<
       InferredBatchParamType
     >();
 
@@ -228,12 +242,23 @@ describe("actions", () => {
       stubData.attachmentUploadRequestBody[stubData.localAttachment1.filename];
 
     const attachment = createAttachmentUpload(blob, "file1.txt");
+
+    // Mimics the Web file API (https://developer.mozilla.org/en-US/docs/Web/API/File). The File constructor is only available in Node 19.2.0 and above
+    const fileAttachment = Object.assign(blob, { name: "file1.txt" });
+
     const result = await client(actionTakesAttachment).applyAction({
       attachment,
     });
 
+    const result2 = await client(actionTakesAttachment).applyAction({
+      attachment: fileAttachment,
+    });
+
     expectTypeOf<typeof result>().toEqualTypeOf<undefined>();
     expect(result).toBeUndefined();
+
+    expectTypeOf<typeof result2>().toEqualTypeOf<undefined>();
+    expect(result2).toBeUndefined();
   });
   it("conditionally returns edits in batch mode", async () => {
     const result = await client(moveOffice).batchApplyAction([