+++ packages/e2e.sandbox.catchall/src/index.ts
@@ -25,6 +25,7 @@ import { runGeotimeSeriesReferenceTests } from "./runGeotimeSeriesTest.js";
 import { runInterfacesTest } from "./runInterfacesTest.js";
 import { runLegacyExamples } from "./runLegacyExamples.js";
 import { runQueriesTest } from "./runQueriesTest.js";
+import { runStructsTest } from "./runStructsTest.js";
 import { runSubscriptionsTest } from "./runSubscriptionsTest.js";
 import { runTimeseriesTest } from "./runTimeseriesTest.js";
 import { typeChecks } from "./typeChecks.js";
@@ -68,6 +69,8 @@ async function runTests() {
     await runTimeseriesTest();
 
     await runGeotimeSeriesReferenceTests();
+
+    await runStructsTest();
   } catch (e) {
     console.error(`Caught an error we did not expect, type: ${typeof e}`);
     console.error(e);