+++ packages/e2e.sandbox.todowidget/src/second.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "@osdk/widget-client.unstable";
 
 const Config = defineConfig({
   type: "workshop",
-  rid: "ri.viewregistry..view.1234-0000-0000-0000",
+  rid: "ri.widgetregistry..widget.1234-0000-0000-0000",
   parameters: {
     headerText: {
       displayName: "Widget title",