+++ packages/shared.test/src/stubs/objectTypesWithLinkTypes.ts
@@ -22,6 +22,7 @@ import {
   peepsLinkType,
 } from "./linkTypes.js";
 import {
+  BGaoNflPlayerObjectType,
   employeeObjectType,
   equipmentObjectType,
   objectTypeWithAllPropertyTypes,
@@ -121,6 +122,14 @@ export const personWithLinkTypes: ObjectTypeFullMetadata = {
   sharedPropertyTypeMapping: {},
 };
 
+export const bGaoNflPlayerWithLinkTypes: ObjectTypeFullMetadata = {
+  objectType: BGaoNflPlayerObjectType,
+  linkTypes: [],
+  implementsInterfaces: [],
+  implementsInterfaces2: {},
+  sharedPropertyTypeMapping: {},
+};
+
 export const objectTypesWithLinkTypes: {
   [objectTypeApiName: string]: ObjectTypeFullMetadata;
 } = {
@@ -135,4 +144,5 @@ export const objectTypesWithLinkTypes: {
   [todoObjectType.apiName]: todoWithLinkTypes,
   [taskObjectType.apiName]: taskWithLinkTypes,
   [personObjectType.apiName]: personWithLinkTypes,
+  [BGaoNflPlayerObjectType.apiName]: bGaoNflPlayerWithLinkTypes,
 };