+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [b773513]
+- Updated dependencies [9d35a1f]
+- Updated dependencies [abfe4b3]
+- Updated dependencies [c537e2d]
+  - @osdk/client@2.1.0-beta.20
+
 ## 0.1.0-beta.2
 
 ### Patch Changes