+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @osdk/generator-converters
 
+## 2.1.0-beta.20
+
+### Minor Changes
+
+- 9d35a1f: Remove unused imports
+- abfe4b3: Add support for reading struct properties.
+- bf0e3b4: Ignore unknown property types.
+
+### Patch Changes
+
+- Updated dependencies [9d35a1f]
+- Updated dependencies [abfe4b3]
+- Updated dependencies [c537e2d]
+  - @osdk/api@2.1.0-beta.20
+
 ## 2.1.0-beta.19
 
 ### Patch Changes