+++ packages/generator-utils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/generator-utils
 
+## 2.1.0-beta.20
+
 ## 2.1.0-beta.19
 
 ## 2.1.0-beta.18