+++ packages/widget.api.unstable/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/widget-api.unstable
 
+## 2.0.0-beta.5
+
 ## 2.0.0-beta.4
 
 ### Major Changes