+++ packages/widget.client-react.unstable/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @osdk/widget-client-react.unstable
 
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- Updated dependencies [b773513]
+- Updated dependencies [9d35a1f]
+- Updated dependencies [abfe4b3]
+- Updated dependencies [c537e2d]
+  - @osdk/client@2.1.0-beta.20
+  - @osdk/widget-client.unstable@2.0.0-beta.5
+
 ## 2.0.0-beta.4
 
 ### Major Changes